# internal/assets/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tumble API",
    "description": "API for the Tumble content application. The v1 API provides a RESTful, versioned interface with consistent request/response formats. Legacy endpoints are deprecated and will be removed in a future release.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for protected endpoints"
      }
    },
    "schemas": {
      "APIError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code",
                "enum": ["bad_request", "unauthorized", "forbidden", "not_found", "validation_error", "internal_error"]
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "description": "Field-level validation errors (for validation_error code)"
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"],
        "example": {
          "error": {
            "code": "not_found",
            "message": "Link 42 does not exist"
          }
        }
      },
...
      "APILinksListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/APILinkResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/APIMeta"
          }
        },
        "required": ["data", "meta"]
      },
...
      "APIQuotesListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/APIQuoteResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/APIMeta"
          }
        },
        "required": ["data", "meta"]
      },
...
      "LinkCreateRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "The URL to submit"
          },
          "user": {
            "type": "string",
            "description": "Username of the submitter"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional tags to add to the link (lowercased, no spaces)"
          }
        },
        "required": ["url", "user"],
        "example": {
          "url": "https://example.com/article",
          "user": "alice",
          "tags": ["golang", "tutorial"]
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/links": {
      "get": {
        "summary": "List Links",
        "description": "Returns a paginated list of links. Can be filtered by user.",
        "tags": ["Links"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results (default: 50, max: 1000)",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip (default: 0)",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "user",
            "in": "query",
            "description": "Filter by submitter username",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of links with pagination metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APILinksListResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "description": "Plain text representation of links"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Link",
        "description": "Submits a new link. Links are always created even if the URL was previously submitted. Duplicate information is included in the response.",
        "tags": ["Links"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Link created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APILinkCreateResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "123",
                  "description": "Link ID (plain text format)"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (malformed JSON)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (invalid URL or missing fields)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/quotes": {
      "get": {
        "summary": "List Quotes",
        "description": "Returns a paginated list of quotes.",
        "tags": ["Quotes"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results (default: 50, max: 1000)",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip (default: 0)",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of quotes with pagination metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIQuotesListResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "description": "Plain text representation of quotes"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Quote",
        "description": "Submits a new quote. Only the quote text is required.",
        "tags": ["Quotes"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quote created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIQuoteResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "42",
                  "description": "Quote ID (plain text format)"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (malformed JSON)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (missing quote text)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stats": {
      "get": {
        "summary": "Get Site Statistics",
        "description": "Returns site-wide statistics and a leaderboard of users by activity.",
        "tags": ["Statistics"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum leaderboard entries (default: 50, max: 1000)",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of leaderboard entries to skip (default: 0)",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Site statistics and leaderboard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIStatsResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "description": "Plain text representation"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/search": {
      "get": {
        "summary": "Search Content",
        "description": "Search links and quotes by keyword. Minimum 4 characters required.",
        "tags": ["Search"],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search query (minimum 4 characters)",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 4
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Content types to search (comma-separated: links, quotes). Defaults to both.",
            "schema": {
              "type": "string",
              "example": "links,quotes"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum results per type (default: 50)",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip per type (default: 0)",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APISearchResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "description": "Plain text representation"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (query too short)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cache": {
      "put": {
        "summary": "Refresh Cached Preview",
        "description": "Deletes the existing cached preview for a URL and re-fetches it. Specify the link by ID or URL.",
        "tags": ["Cache"],
        "security": [
          { "apiKey": [] }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Link ID to refresh the cached preview for",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "url",
            "in": "query",
            "description": "URL to refresh the cached preview for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Preview refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APICacheRefreshResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing id/url parameter)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (invalid API key)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Link not found (when using id parameter)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Clear Preview Cache",
        "description": "Clears the link preview cache. Can clear all entries or a specific URL.",
        "tags": ["Cache"],
        "security": [
          { "apiKey": [] }
        ],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "description": "Specific URL to clear from cache. If omitted, clears all cache.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cache cleared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APICacheResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (missing API key)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (invalid API key)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/caching/invalidate": {
      "get": {
        "summary": "Invalidate Link Preview Cache (Deprecated)",
        "description": "DEPRECATED: Use DELETE /api/v1/cache instead. Removes a URL from the server-side link preview cache.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "description": "The URL to invalidate",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cache Invalidated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing URL"
          }
        }
      }
    },
    "/api/kitten/fetch": {
      "post": {
        "summary": "Force Fetch Daily Kitten (Deprecated)",
        "description": "DEPRECATED: Use PUT /api/v1/kittens/daily instead. Forces a new daily kitten to be fetched from cataas.com.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "responses": {
          "200": {
            "description": "Kitten fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "url": { "type": "string", "example": "https://cataas.com/cat/abc123xyz" }
                  }
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed (use POST)"
          },
          "500": {
            "description": "Failed to fetch kitten",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "error" },
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/link/": {
      "post": {
        "summary": "Submit a new Link (Deprecated)",
        "description": "DEPRECATED: Use POST /api/v1/links instead. Submits a new link using query parameters.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "description": "The username of the submitter",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "url",
            "in": "query",
            "description": "The URL to submit",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Source of submission",
            "schema": {
              "type": "string",
              "enum": ["irc", "api", "web"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Link created"
          },
          "208": {
            "description": "Duplicate link created"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
    "/quote/": {
      "get": {
        "summary": "Get a Random Quote (Deprecated)",
        "description": "DEPRECATED: Use GET /api/v1/quotes instead.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "responses": {
          "200": {
            "description": "Random Quote"
          }
        }
      },
      "post": {
        "summary": "Submit a Quote (Deprecated)",
        "description": "DEPRECATED: Use POST /api/v1/quotes instead.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "quote": { "type": "string" },
                  "author": { "type": "string" },
                  "poster": { "type": "string" }
                },
                "required": ["quote"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quote created"
          },
          "400": {
            "description": "Missing inputs"
          }
        }
      }
    },
...
    "/search": {
      "get": {
        "summary": "Search Content (Deprecated)",
        "description": "DEPRECATED: Use GET /api/v1/search instead.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 4
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search Results HTML"
          }
        }
      }
    },
...
    "/stats.json": {
      "get": {
        "summary": "Get User Statistics as JSON (Deprecated)",
        "description": "DEPRECATED: Use GET /api/v1/stats instead.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of user statistics"
          }
        }
      }
    },
    "/ogpreview": {
      "get": {
        "summary": "Get OpenGraph Preview (Deprecated)",
        "description": "DEPRECATED: This endpoint may be removed in a future release.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OpenGraph Metadata"
          }
        }
      }
    }
  },
...
}